`DefaultCompositionProvider` extends `CompositionProvider`
